chore: move inline test harnesses to tests/ directories
- product_computation.rs: 3 #[cfg(test)] modules (kernel oracle, recording
  FSProver equality, lazy-vs-eager transcript) → tests/product_sumcheck_equality.rs
- poseidon/mod.rs: witness microbench + deferred_aux_fill + packed_witness
  → tests/poseidon_witness.rs (mds_u128_equals_oracle stays inline — needs
  private mds_vec_mul)
